Cube Placement Corrected
Cubes can now be placed correctly on the grid. Multiple Cubes can still be stacked in the same spot, but it's not necessary to fix as it is an exercise for the grid.

Work on hovering ghost tiles next (Shows player where they would place a tile on the grid.)